Refactor plans, stubs to be consistent with expectations
