+++ src/i22_bluesky/plans/__init__.py
@@ -1,20 +1,17 @@
-from .linkam import linkam_plan, save_linkam
+from .common import check_devices
+from .linkam import linkam_plan
 from .stopflow import (
-    check_detectors_for_stopflow,
     check_stopflow_assembly,
     check_stopflow_experiment,
-    save_stopflow,
     stopflow,
     stress_test_stopflow,
 )
 
 __all__ = [
+    "check_devices",
     "linkam_plan",
-    "save_linkam",
     "stopflow",
-    "check_detectors_for_stopflow",
     "check_stopflow_assembly",
     "check_stopflow_experiment",
-    "save_stopflow",
     "stress_test_stopflow",
 ]
