+++ src/i22_bluesky/plans/stopflow.py
@@ -19,62 +19,35 @@
 
 from typing import Any
 
-import bluesky.plan_stubs as bps
-import bluesky.plans as bp
 import bluesky.preprocessors as bpp
 from bluesky.protocols import Readable
 from bluesky.utils import MsgGenerator
-from dodal.devices.tetramm import TetrammDetector
 from dodal.plans.data_session_metadata import attach_data_session_metadata_decorator
 from ophyd_async.core import (
-    DetectorTrigger,
     StandardDetector,
     StandardFlyer,
-    TriggerInfo,
-    in_micros,
-    load_device,
-    save_device,
 )
 from ophyd_async.fastcs.panda import (
     HDFPanda,
-    SeqTable,
-    SeqTableInfo,
-    SeqTrigger,
     StaticSeqTableTriggerLogic,
 )
-from ophyd_async.plan_stubs import fly_and_collect
+from ophyd_async.plan_stubs import ensure_connected, fly_and_collect
 
+from i22_bluesky.stubs.panda import load_panda_config_for_stopflow
+from i22_bluesky.stubs.stopflow import (
+    prepare_seq_table_flyer_and_det,
+    raise_for_minimum_exposure_times,
+)
 from i22_bluesky.util.default_devices import (
     BASELINE_DEVICES,
     DETECTORS,
     FAST_DETECTORS,
     PANDA,
 )
-from i22_bluesky.util.settings import get_device_save_dir
 
 #: Buffer added to deadtime to handle minor discrepencies between detector
 #: and panda clocks
 DEADTIME_BUFFER = 20e-6
-
-
-@attach_data_session_metadata_decorator()
-def check_detectors_for_stopflow(
-    num_frames: int = 1,
-    devices: set[Readable] = DETECTORS | BASELINE_DEVICES,
-) -> MsgGenerator:
-    """
-    Take a reading from all devices that are used in the
-    stopflow plan by default
-    """
-
-    # Tetramms do not support software triggering
-    software_triggerable_devices = {
-        device for device in devices if not isinstance(device, TetrammDetector)
-    }
-    yield from bp.count(
-        tuple(software_triggerable_devices),
-        num=num_frames,
-    )
 
 
 def check_stopflow_assembly(
@@ -87,7 +60,6 @@ def check_stopflow_assembly(
     wiring/assembly/detector setup errors, does not require triggering a stop flow,
     that can be tested with the main plan.
     """
-
     yield from stopflow(
         exposure=0.1,
         post_stop_frames=0,
@@ -138,14 +110,6 @@ def stress_test_stopflow(
     )
 
 
-def save_stopflow(panda: HDFPanda = PANDA) -> MsgGenerator:
-    yield from save_device(
-        panda,
-        get_device_save_dir(stopflow.__name__),
-        ignore=["pcap.capture", "data.capture", "data.datasets"],
-    )
-
-
 def stopflow(
     exposure: float,
     post_stop_frames: int,
@@ -179,6 +143,7 @@ def stopflow(
     Yields:
             Iterator[MsgGenerator]: Bluesky messages
     """
+    yield from ensure_connected(panda, *detectors, *baseline)
 
     # Check that all detectors supplied can actually go as
     # fast as requested
@@ -213,7 +178,7 @@ def stopflow(
     @bpp.stage_decorator(devices)
     @bpp.run_decorator(md=_md)
     def inner_stopflow_plan():
-        yield from load_device(panda, get_device_save_dir(stopflow.__name__))
+        yield from load_panda_config_for_stopflow(panda)
         yield from prepare_seq_table_flyer_and_det(
             flyer=flyer,
             detectors=detectors,
@@ -229,162 +194,3 @@ def stopflow(
         )
 
     yield from inner_stopflow_plan()
-
-
-def prepare_seq_table_flyer_and_det(
-    flyer: StandardFlyer[SeqTableInfo],
-    detectors: set[StandardDetector],
-    pre_stop_frames: int,
-    post_stop_frames: int,
-    exposure: float,
-    shutter_time: float,
-    period: float = 0.0,
-) -> MsgGenerator:
-    """
-    Setup detectors/flyer for a stop flow experiment. Create a seq table and
-    upload it to the panda. Arm all detectors.
-
-    Args:
-            flyer: Flyer object that controls the panda
-            detectors: Detectors that are triggered by the panda
-            post_stop_frames: Number of frames to be collected after the flow
-                    is stopped.
-            pre_stop_frames: Number of frames (if any) to be collected before
-                    the flow is stopped.
-            exposure: Detector exposure time
-            shutter_time: Time period (seconds) to wait for the shutter to
-                    open fully before beginning acquisition
-            period: Time period (seconds) to wait after arming the detector
-                    before taking the first batch of frames
-
-    Returns:
-            MsgGenerator: Plan
-
-    Yields:
-            Iterator[MsgGenerator]: Bluesky messages
-    """
-
-    deadtime = (
-        max(det.controller.get_deadtime(exposure) for det in detectors)
-        + DEADTIME_BUFFER
-    )
-    trigger_info = TriggerInfo(
-        num=(pre_stop_frames + post_stop_frames),
-        trigger=DetectorTrigger.constant_gate,
-        deadtime=deadtime,
-        livetime=exposure,
-        frame_timeout=60.0,
-    )
-
-    # Generate a seq table
-    table = stopflow_seq_table(
-        pre_stop_frames,
-        post_stop_frames,
-        exposure,
-        shutter_time,
-        deadtime,
-        period,
-    )
-    table_info = SeqTableInfo(table, repeats=1)
-
-    # Upload the seq table and arm all detectors.
-    for det in detectors:
-        yield from bps.prepare(det, trigger_info, wait=False, group="prep")
-    yield from bps.prepare(flyer, table_info, wait=False, group="prep")
-    yield from bps.wait(group="prep")
-
-
-def stopflow_seq_table(
-    pre_stop_frames: int,
-    post_stop_frames: int,
-    exposure: float,
-    shutter_time: float,
-    deadtime: float,
-    period: float,
-) -> SeqTable:
-    """Create a SeqTable based on the parameters of a stop flow measurement
-
-    Args:
-            pre_stop_frames: Number of frames to take initially, before flow stops
-            post_stop_frames: Number of frames to take after flow stops
-            exposure: Exposure time of each frame (excluding deadtime)
-            shutter_time: Time period (seconds) to wait for the shutter
-                    to open fully before beginning acquisition
-            deadtime: Dead time to leave between frames, dependant on the
-                    instruments involved
-            period: Time period (seconds) to wait after arming the detector
-                    before taking the first batch of frames
-
-    Returns:
-            SeqTable: SeqTable that will result in a series of triggers
-                    for the measurement
-    """
-
-    total_gate_time = (pre_stop_frames + post_stop_frames) * (exposure + deadtime)
-    pre_delay = max(period - 2 * shutter_time - total_gate_time, 0)
-    # Wait for pre-delay then open shutter
-    table = SeqTable.row(
-        time1=in_micros(pre_delay),
-        time2=in_micros(shutter_time),
-        outa2=True,
-    )
-
-    # Keeping shutter open, do n triggers
-    if pre_stop_frames > 0:
-        table += SeqTable.row(
-            repeats=pre_stop_frames,
-            time1=in_micros(exposure),
-            outa1=True,
-            outb1=True,
-            time2=in_micros(deadtime),
-            outa2=True,
-        )
-    # Do m triggers after BITA=1
-    if post_stop_frames > 0:
-        table += SeqTable.row(
-            trigger=SeqTrigger.BITA_1,
-            repeats=1,
-            time1=in_micros(exposure),
-            outa1=True,
-            outb1=True,
-            time2=in_micros(deadtime),
-            outa2=True,
-        )
-        if post_stop_frames > 1:
-            table += SeqTable.row(
-                repeats=post_stop_frames - 1,
-                time1=in_micros(exposure),
-                outa1=True,
-                outb1=True,
-                time2=in_micros(deadtime),
-                outa2=True,
-            )
-    # Add the shutter close
-    table += SeqTable.row(time2=in_micros(shutter_time))
-    return table
-
-
-def raise_for_minimum_exposure_times(
-    exposure: float,
-    detectors: set[StandardDetector],
-) -> None:
-    minimum_exposure_times = {
-        "saxs": 1.0 / 250.0,
-        "waxs": 1.0 / 250.0,
-        "oav": 1.0 / 22.0,
-        "i0": 1.0 / 2e4,
-        "it": 1.0 / 2e4,
-    }
-    detectors_below_limit = {
-        detector
-        for detector in detectors
-        if exposure < minimum_exposure_times.get(detector.name, 0.0)
-    }
-    if len(detectors_below_limit) > 0:
-        raise KeyError(
-            f"The exposure time requested was {exposure}, but "
-            "the following detectors do not support going "
-            f"that fast: {detectors_below_limit}. Try running the plan"
-            "without them. "
-            f"See minimum exposure time table: {minimum_exposure_times}"
-        )
