--- src/i22_bluesky/stubs/__init__.py
@@ -1,13 +1,0 @@
-from .linkam import (
-    LinkamPathSegment,
-    LinkamTrajectory,
-    capture_linkam_segment,
-    capture_temp,
-)
-
-__all__ = [
-    "LinkamPathSegment",
-    "LinkamTrajectory",
-    "capture_linkam_segment",
-    "capture_temp",
-]
